+++ server/src/controllers/kyc/getAllKyc.controller.js
@@ -1,9 +1,7 @@
-import ApiError from "../../utils/ApiError.js";
-import ApiResponse from "../../utils/ApiResponse.js";
-import asyncHandler from "../../utils/asyncHandler.js";
+import { ApiResponse, asyncHandler } from "../../utils/index.js";
 
-export const getAllKyc = asyncHandler(async (req, res) => {
+export const getAllKyc = asyncHandler(async (_, res) => {
     return res
         .status(200)
-        .json(new ApiResponse(200, true, true, "test passed", null));
+        .json(new ApiResponse(200, true, true, "Test passed", null));
 });
