+++ server/src/controllers/user/user.controller.js
@@ -1,9 +1,12 @@
-import asyncHandler from "../utils/asyncHandler.js";
-import ApiResponse from "../utils/ApiResponse.js";
-import { User } from "../models/user.model.js";
-import { MailService } from "../utils/MailHandler.js";
 import jwt from "jsonwebtoken";
-import ApiError from "../utils/ApiError.js";
+import { User } from "../../models/index.js";
+import {
+    ApiResponse,
+    ApiError,
+    asyncHandler,
+    MailService,
+} from "../../utils/index.js";
+
 
 export const generateAccessToken = async function (id) {
     const user = await User.findById(id);
