+++ server/src/models/index.js
@@ -1,0 +1,5 @@
+import { User } from "./user.model.js";
+import { Kyc } from "./kyc.model.js";
+import { Otp } from "./otp.model.js";
+
+export { User, Kyc, Otp };
